Move API token functions to its own file
